chore: Move api folder to lib
