+++ apps/editor.planx.uk/src/@planx/components/List/Public/tests/index.listBuilding.test.tsx
@@ -1,5 +1,5 @@
 import { within } from "@testing-library/react";
-import { uploadPrivateFile } from "api/fileUpload/requests";
+import { uploadPrivateFile } from "lib/api/upload";
 import { cloneDeep, merge } from "lodash";
 import React from "react";
 import { setup } from "testUtils";
