+++ apps/editor.planx.uk/src/@planx/components/List/Public/tests/index.payloadGeneration.test.tsx
@@ -1,5 +1,5 @@
 import { screen } from "@testing-library/react";
-import { uploadPrivateFile } from "api/fileUpload/requests";
+import { uploadPrivateFile } from "lib/api/upload";
 import React from "react";
 import { setup } from "testUtils";
 import { it, vi } from "vitest";
