+++ apps/editor.planx.uk/src/lib/api/downloadApplicationFiles/requests.tsx
@@ -1,4 +1,4 @@
-import apiClient from "api/client";
+import apiClient from "lib/api/client";
 
 interface DownloadApplicationFiles {
   sessionId: string;
