+++ apps/editor.planx.uk/src/lib/api/send/requests.ts
@@ -1,4 +1,4 @@
-import apiClient from "api/client";
+import apiClient from "lib/api/client";
 
 import { CombinedEventsPayload, SendResponse } from "./types";
