+++ apps/editor.planx.uk/src/lib/api/slack/requests.ts
@@ -1,4 +1,4 @@
-import apiClient from "api/client";
+import apiClient from "lib/api/client";
 
 export const sendSlackMessage = async (message: string) =>
   await apiClient.post("/send-slack-notification", { message });
