+++ apps/editor.planx.uk/src/pages/FlowEditor/components/Sidebar/EditHistory/TimelineItems.tsx
@@ -3,7 +3,7 @@ import { styled } from "@mui/material/styles";
 import Typography from "@mui/material/Typography";
 import SimpleExpand from "@planx/components/shared/Preview/SimpleExpand";
 import { formatOps } from "@planx/graph";
-import { CommentHistoryItem, OperationHistoryItem, PublishHistoryItem } from "api/publishFlow/types";
+import { CommentHistoryItem, OperationHistoryItem, PublishHistoryItem } from "lib/api/publishFlow/types";
 import { Store } from "pages/FlowEditor/lib/store";
 import React from "react";
 import BlockQuote from "ui/editor/BlockQuote";
