+++ apps/editor.planx.uk/src/pages/FlowEditor/components/Sidebar/EditHistory/index.tsx
@@ -2,8 +2,8 @@ import { gql, useSubscription } from "@apollo/client";
 import Box from "@mui/material/Box";
 import Divider from "@mui/material/Divider";
 import Typography from "@mui/material/Typography";
-import { HistoryItem } from "api/publishFlow/types";
 import DelayedLoadingIndicator from "components/DelayedLoadingIndicator/DelayedLoadingIndicator";
+import { HistoryItem } from "lib/api/publishFlow/types";
 import { useStore } from "pages/FlowEditor/lib/store";
 import React from "react";
