+++ apps/editor.planx.uk/src/pages/FlowEditor/components/Sidebar/Publish/AlteredNodes.tsx
@@ -6,7 +6,7 @@ import ListItemText from "@mui/material/ListItemText";
 import { styled } from "@mui/material/styles";
 import Typography from "@mui/material/Typography";
 import { ComponentType as TYPES } from "@opensystemslab/planx-core/types";
-import { AlteredNode, HistoryItem } from "api/publishFlow/types";
+import { AlteredNode, HistoryItem } from "lib/api/publishFlow/types";
 import { useStore } from "pages/FlowEditor/lib/store";
 import { formatLastEditDate } from "pages/FlowEditor/utils";
 import React, { useState } from "react";
