+++ apps/editor.planx.uk/src/pages/FlowEditor/components/Sidebar/Publish/hooks/usePublishFlow.tsx
@@ -1,8 +1,8 @@
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { logger } from "airbrake";
-import { checkForChanges, publishFlow } from "api/publishFlow/requests";
-import { PublishFlowArgs } from "api/publishFlow/types";
 import { AxiosError } from "axios";
+import { checkForChanges, publishFlow } from "lib/api/publishFlow/requests";
+import { PublishFlowArgs } from "lib/api/publishFlow/types";
 import { useStore } from "pages/FlowEditor/lib/store";
 import { formatLastPublishMessage } from "pages/FlowEditor/utils";
