+++ apps/editor.planx.uk/src/pages/SubmissionDownload/VerifySubmissionEmail.tsx
@@ -5,9 +5,9 @@ import Card from "@planx/components/shared/Preview/Card";
 import { CardHeader } from "@planx/components/shared/Preview/CardHeader/CardHeader";
 import { SummaryListTable } from "@planx/components/shared/Preview/SummaryList";
 import { useMutation } from "@tanstack/react-query";
-import { downloadApplicationFiles } from "api/downloadApplicationFiles/requests";
 import DelayedLoadingIndicator from "components/DelayedLoadingIndicator/DelayedLoadingIndicator";
 import { useFormik } from "formik";
+import { downloadApplicationFiles } from "lib/api/downloadApplicationFiles/requests";
 import startCase from "lodash/startCase.js";
 import React from "react";
 import InputLabel from "ui/public/InputLabel";
