+++ apps/editor.planx.uk/src/pages/Team/components/MoveDialog.tsx
@@ -9,10 +9,10 @@ import MenuItem from "@mui/material/MenuItem";
 import Skeleton from "@mui/material/Skeleton";
 import { Team } from "@opensystemslab/planx-core/types";
 import { useMutation } from "@tanstack/react-query";
-import { moveFlow } from "api/flow/requests";
 import { isAxiosError } from "axios";
 import { Form, Formik, FormikConfig } from "formik";
 import { useToast } from "hooks/useToast";
+import { moveFlow } from "lib/api/flow/requests";
 import { useStore } from "pages/FlowEditor/lib/store";
 import React from "react";
 import SelectInput from "ui/editor/SelectInput/SelectInput";
