+++ apps/editor.planx.uk/src/ui/shared/PublicFileUploadButton.tsx
@@ -4,7 +4,7 @@ import ButtonBase, { ButtonBaseProps } from "@mui/material/ButtonBase";
 import CircularProgress from "@mui/material/CircularProgress";
 import { styled } from "@mui/material/styles";
 import Tooltip from "@mui/material/Tooltip";
-import { uploadPublicFile } from "api/fileUpload/requests";
+import { uploadPublicFile } from "lib/api/upload";
 import React, { useCallback, useEffect, useState } from "react";
 import { FileWithPath, useDropzone } from "react-dropzone";
